Lay ground work for enemy and combat boundaries
Begin toying with basic enemy wandering AI as well as aggro radius and combat triggers. Started working on a way for the game to check where the nearest rectangular space that is free to setup combat would be and then moving the player and the enemy across from each other in said rectangle while entering turn-based mode.

Did a lot of file organizing.