Setup enemy attack selection
Worked on a system using a float->attackType map that allows for the enemy AI to make a weighted semi-random choice between its attack moveset. Also began basic spaghetti code setup for the player and enemy taking their individual turns and got a very rough draft of how I want it to work.